Convert EuiCopy to TypeScript
