+++ src/components/copy/index.ts
